Moved test utils into separate file, cleaned up repo, renamed README to SETUP
