+++ admin-events/agent-config.json
@@ -1,18 +1,34 @@
 {
-  "developerAbbreviation": "AE",
-  "protocolName": "UNISWAP",
-  "protocolAbbreviation": "UNI",
-  "contracts": {
-    "GovernorBravo": {
-      "address": "0x408ED6354d4973f66138C91495F2f2FCbd8724C3",
-      "abiFile": "GovernorBravo.json",
+  "developerAbbreviation": "",
+  "protocolName": "",
+  "protocolAbbreviation": "",
+  "adminEvents": {
+    "contractName1": {
+      "address": "contractAddress1",
+      "abiFile": "filename1.json",
       "events": {
-        "ProposalCreated": {
-          "expression": "proposer !== 0x9B68c14e936104e9a7a24c712BEecdc220002984",
-          "type": "Info",
-          "severity": "Info"
+        "EventName": {
+          "expression": "address !== 0x0000000000000000000000000000000000000000",
+          "expectedResult": true,
+          "type": "Type",
+          "severity": "Severity"
         }
       }
+    },
+    "contractName2": {
+      "address": "contractAddress2",
+      "abiFile": "filename2.json",
+      "events": {
+        "EventName1": {
+          "type": "Type",
+          "severity": "Severity"
+        }
+      }
+    },
+    "contractName3": {
+      "address": "contractAddress3",
+      "abiFile": "filename3.json",
+      "proxy": "contractName2"
     }
   }
 }
